refactor: common to core
