+++ .github/workflows/run_tests.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: crates/yaxp-common
+        working-directory: crates/yaxp-core
     steps:
       - name: Checkout code
         uses: actions/checkout@v3
