+++ crates/yaxp-cli/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "yaxp-cli"
-version = "0.1.16"
+version = "0.1.17"
 edition = "2021"
 authors = ["Jeroen <jeroen@flexworks.eu>"]
 description = "<yaxp-cli ⚡> Yet Another XML Parser CLI"
@@ -10,7 +10,7 @@ license = "MIT"
 
 [dependencies]
 clap = { version = "4.5.23", features = ["derive"] }
-yaxp-common = { path = "../yaxp-common", version = "0.1.16" }
+yaxp-core = { path = "../yaxp-core", version = "0.1.0" }
 serde = { version = "1.0.217", features = ["derive"] }
 encoding_rs = "0.8.35"
 serde_json = "1.0.138"
