+++ crates/yaxp-core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "yaxp-common"
-version = "0.1.17"
+name = "yaxp-core"
+version = "0.1.0"
 edition = "2021"
 authors = ["Jeroen <jeroen@flexworks.eu>"]
 description = "<yaxp-cli ⚡> Yet Another XML Parser"
