+++ crates/yaxp-core/src/lib.rs
@@ -1,0 +1,2 @@
+// mod tests;
+pub mod xsdp;
